requirements.txt
requests
eventlet
flask
flask-socketio
onvif-zeep-async